Paiement_Service/requirements.txt
flask
flask-restx
flask-pymongo
flask-cors
python-dotenv
py-eureka-client
pyjwt